Fix most review comments.
