+++ tao_compiler/mlir/disc/disc_compiler.cc
@@ -294,9 +294,10 @@ LogicalResult LowerHLOToLLVM(ModuleOp m, const DISCLoweringOptions& options) {
   if (enable_stitch && gpu_enabled) {
     // Some passes introduce a bunch of memref alloc, load and store to for
     // shape operations (e.g., ReshapeOp(.., target_shape, ..)), which makes
-    // shape equality analysis quite tricky. A GVN pass helps to eliminate
+    // shape equality analysis quite tricky. This pass helps to eliminate
     // unnecessary shape value transformation.
-    pm.addNestedPass<FuncOp>(disc_ral::createDiscMemRefGVNPass());
+    pm.addNestedPass<FuncOp>(
+        disc_ral::createDiscMemRefLoadStoreSimplifierPass());
   }
   pm.addNestedPass<FuncOp>(disc_ral::createDiscFusionPass(
       gpu_enabled, enable_stitch ? "stitch" : "base"));
@@ -331,8 +332,9 @@ LogicalResult LowerHLOToLLVM(ModuleOp m, const DISCLoweringOptions& options) {
   if (enable_stitch && gpu_enabled) {
     // The passes between `DiscLhloLegalizeRootsToParallelLoops` and
     // 'DiscFusionPass' introduce new shape value transformations with extra
-    // memref load ans store. Eliminate them with GVN.
-    pm.addNestedPass<FuncOp>(disc_ral::createDiscMemRefGVNPass());
+    // memref load and store. Eliminate them.
+    pm.addNestedPass<FuncOp>(
+        disc_ral::createDiscMemRefLoadStoreSimplifierPass());
   }
   // CodeGen passes: lhlo -> gpu.launch_func
   // TODO: move to aicompiler repo and add more schedules/op coverage
